Roll a custom VCF parser
The bionode-vcf package has a number of bugs that are a real pain. The
most troubling is that it doesn't handle mulitple-sample VCF files
correctly. In particular, it uses the same data for all samples. While
it would be neighborly to submit issues and bug fixes, we don't really
have time for that. Insted, we decided to just roll our own.