Fix test indeterminacy
The `geneSorter` function was indeterminate when two regions had
the same `Start` and `End` values, e.g. when a gene codes for an
mRNA. The fix is to also sort by ID/ParentID. If you two entries
have the same range and one is the parent of the other, then the
parent should come first. This doesn't necessarily clear up all of
indeterminacy, but it's good enough.